moved interaction_dataset_reader to seperate module
