+++ examples/BUILD
@@ -154,11 +154,12 @@ py_binary(
         ":params",
         "//python:bark.so",
     ] + 
-    select({"//modules/runtime/scenario/dataset:_interaction_dataset_present" : ["@interaction_dataset_fortiss_internal//:maps_tracks"], "//conditions:default": []}),
+    select({"//modules/runtime/scenario/interaction_dataset_processing:_interaction_dataset_present" : ["@interaction_dataset_fortiss_internal//:maps_tracks"], "//conditions:default": []}),
     imports = ["../python/"],
     deps = [
         "//modules/runtime/commons:parameters",
-        "//modules/runtime/scenario/scenario_generation",
+        "//modules/runtime/scenario/scenario_generation:scenario_generation",
+        "//modules/runtime/scenario/interaction_dataset_processing:interaction_dataset_processing",
         "//modules/runtime/viewer:matplotlib_viewer",
         "//modules/runtime/viewer:video_renderer",
         "//modules/runtime:runtime",
