+++ modules/runtime/scenario/scenario_generation/BUILD
@@ -3,12 +3,10 @@ py_library(
     srcs = ["scenario_generation.py",
             "uniform_vehicle_distribution.py",
             "configurable_scenario_generation.py",
-            "model_json_conversion.py",
             "deterministic.py",
             "config_with_ease.py",
             "deterministic_drone_challenge.py",
             "drone_challenge.py",
-            "interaction_dataset_reader.py",
             "interaction_dataset_scenario_generation.py",
             "interaction_dataset_scenario_generation_full.py"],
     data = ['//python:bark.so'],
@@ -17,7 +15,8 @@ py_library(
         "//modules/runtime/scenario:scenario",
         "@com_github_interaction_dataset_interaction_dataset//:interaction-tools",
         "//modules/runtime/scenario/scenario_generation/config_readers:config_readers",
-        "//modules/runtime/scenario/dataset:dataset_decomposer",
+        "//modules/runtime/scenario/interaction_dataset_processing:interaction_dataset_processing",
+        "//modules/runtime/commons:model_json_conversion",
     ],
     visibility = ["//visibility:public"],
 )
