+++ modules/runtime/tests/py_interaction_dataset_reader_test.py
@@ -4,7 +4,7 @@
 # https://opensource.org/licenses/MIT
 
 import unittest
-from modules.runtime.scenario.scenario_generation.interaction_dataset_reader import agent_from_trackfile
+from modules.runtime.scenario.interaction_dataset_processing.interaction_dataset_reader import agent_from_trackfile
 from modules.runtime.commons.parameters import ParameterServer
 
 class InteractionDatasetReaderTest(unittest.TestCase):
